Solved single-levels test issue
Added an #if UNITY_EDITOR condition that hard sets the value of the level counter when single level testing, instead of getting the value from the game manager. Indeed, in single level test the Game Manager does not store the current level number, causing the affiliated value to be null.